enhance: add latest revision to write options
If the latest revision write option is passed and this is not the latest
revision, then writing the file will fail with a conflict error. If the
latest revision write option is not passed, then no check is done and
writing will not fail with a conflict.

+++ pkg/client/fileinfo.go
@@ -16,5 +16,5 @@ type RevisionInfo struct {
 }
 
 type revisionInfo struct {
-	CurrentID int `json:"currentID"`
+	CurrentID int64 `json:"currentID"`
 }
